Added support for new board rev and new battery ic
rfcx-battery.c and rfcx-battery.h were mainly
changed, giving support for the mirochip mcp73871
battery charge management IC. The main changes is that
there are more status inputs. Also this is a new rev of board
so some pins changed. I really do not like having PORT,PIN,or DDR macros
anywhere but the header files where other macros are used to point to
them, making it easy for a user to change. It is very bad form to have
those macros in functions. It is a nightmare to change.
The rfcx-mcu.c was changed so that the rfcx_batteries_init is not in the
port init function because otherwise the batteries object has to be
passed around now for that.
The .gitignore was changed so that it does not take in
"compile_commands.json" because I've been using "bear" to make clang
json build commands files for emacs to parse the source files